+++ src/tokens/components/error-page.ts
@@ -1,4 +1,0 @@
-import errorPage from '../../../tokens/component/error-page.json';
-import { flattenTokens } from '../utils';
-
-export const errorPageTokens = flattenTokens(errorPage as never);