+++ src/tokens/components/inline.ts
@@ -1,0 +1,4 @@
+import inlineJson from '../../../tokens/component/inline.json';
+import { flattenTokens } from '../utils';
+
+export const inlineTokens = flattenTokens(inlineJson as never);